+++ db/setup.sql
@@ -7,9 +7,12 @@ CREATE TABLE IF NOT EXISTS `users` (
 
 INSERT IGNORE INTO `users` (`username`, `password`, `enabled`) VALUES
         ('alper', '{noop}123', 1),
-        ('john', '{bcrypt}$2y$12$2uyIQbVvq2STqGm8BZZpy.NHqm2KHyH8OrItzqDXk.AsiuPdu.D4O', 1),
-        ('mary', '{bcrypt}$2a$04$eFytJDGtjbThXa80FyOOBuFdK2IwjyWefYkMpiBEFlpBwDH.5PM0K', 1),
-        ('susan', '{noop}test123', 1);
+        ('gonenc', '{noop}123', 1),
+        ('tcan', '{noop}123', 1),
+        ('kubilay', '{noop}123', 1),
+        ('ali', '{noop}123', 1),
+        ('sadik', '{noop}123', 1),
+        ('adem', '{noop}123', 1);
 
 CREATE TABLE IF NOT EXISTS `authorities` (
   `username` varchar(50) NOT NULL,
@@ -21,7 +24,9 @@ CREATE TABLE IF NOT EXISTS `authorities` (
 INSERT IGNORE INTO `authorities` (`username`, `authority`) VALUES
         ('alper', 'ROLE_ADMIN'),
         ('alper', 'ROLE_USER'),
-        ('john', 'ROLE_USER'),
-        ('mary', 'ROLE_USER'),
-        ('susan', 'ROLE_ADMIN'),
-        ('susan', 'ROLE_USER');
+        ('gonenc', 'ROLE_USER'),
+        ('tcan', 'ROLE_USER'),
+        ('kubilay', 'ROLE_USER'),
+        ('ali', 'ROLE_USER'),
+        ('sadik', 'ROLE_USER'),
+        ('adem', 'ROLE_USER');