+++ src/main/java/com/alper/worldcup/service/PointsService.java
@@ -1,6 +1,8 @@
 package com.alper.worldcup.service;
 
+import com.alper.worldcup.entity.MatchStage;
+
 public interface PointsService {
 
-    int calculatePoints(int guessHome, int guessAway, int actualHome, int actualAway);
+    int calculatePoints(int guessHome, int guessAway, int actualHome, int actualAway, MatchStage stage);
 }